Moved :truck: crates/db -> src/
Issue: #41

squash
