--- crates/db/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "db"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-name = "db"
-path = "core/lib.rs"
-
-[dependencies]
-entity = { path = "entity" }
-migration = { path = "migration" }
-
-chrono = "0.4.24"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
