--- crates/db/entity/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "entity"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-serde = "1.0.162"
-
-[dependencies.sea-orm]
-version = "0.11.3"
-features = [
-    "debug-print",
-    "runtime-async-std-native-tls",
-    "sqlx-postgres",
-]
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
