--- crates/db/migration/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "migration"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies.sea-orm-migration]
-version = "0.11.0"
-features = [
-    "runtime-async-std-native-tls",
-    "sqlx-postgres",
-]
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
