+++ src/db/crud.rs
@@ -1,0 +1,2 @@
+pub mod mutation;
+pub mod query;
