+++ src/db/entities.rs
@@ -4,6 +4,3 @@ pub mod prelude;
 
 pub mod comment;
 pub mod post;
-
-pub use sea_orm;
-pub use serde;
