+++ src/db/migrator.rs
@@ -3,8 +3,6 @@ use sea_orm_migration::prelude::*;
 mod m20230501_000001_create_post_table;
 mod m20230501_000002_create_comment_table;
 
-pub use sea_orm_migration;
-
 pub struct Migrator;
 
 #[async_trait::async_trait]
