+++ src/routes.rs
@@ -5,7 +5,7 @@ use serde_json::to_string_pretty;
 
 use std::env;
 
-// mod posts;
+mod posts;
 
 #[derive(Serialize)]
 #[serde(crate = "self::serde")]
@@ -29,5 +29,5 @@ async fn index() -> HttpResponse {
 pub fn init(cfg: &mut web::ServiceConfig) {
     cfg.service(index);
 
-    // cfg.configure(posts::init_routes);
+    cfg.configure(posts::init_routes);
 }
